+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Make component auditing more resilient ([PR #1604](https://github.com/alphagov/govuk_publishing_components/pull/1604))
 * Add more font sizes to heading component ([PR #1587](https://github.com/alphagov/govuk_publishing_components/pull/1587))
 * Format machine-readable output ([PR #1617](https://github.com/alphagov/govuk_publishing_components/pull/1617))
+* Add support for custom sizes for radio page headers ([PR #1616](https://github.com/alphagov/govuk_publishing_components/pull/1616))
 
 ## 21.59.0